Control class for FLSepVPA

Objects of this class contain all the neccessary settings for the Separable VPA model available through function SepVPA.

Objects from the Class

Objects can be created by calls of the form new("FLSepVPA.control", ...) or by calling the FLSepVPA.control function.

Slots

sep.nyr:

Object of class "integer". Number of years for separable model.

sep.age:

Object of class "integer". Reference age for fitting the separable model. Default value = 4

sep.sel:

Object of class "numeric". Default value = 1.0

Methods

No methods defined with class "FLSepVPA.control" in the signature.

References

JG Shepherd, SM Stevens. 1983. Separable VPA: User's guide - Int. Rep., MAFF Direct. Fish. Res.

See also

Examples

sep.vpa.control <- FLSepVPA.control(sep.nyr=5, sep.age=5)