Projection preparation for FLR

Extends an FLStock object along the year dimensin and prepares some of the slots for forward stock projection, using a similar method to FLSTF. Can be used in conjunction with fwd() for validating assessments.

Generic Function

Usage

stf(object,nyears=3,wts.nyears=3,fbar.nyears=NA,arith.mn=TRUE, na.rm=TRUE)

Arguments:

\

Value

FLStock object

Methods

object="FLStock"

: Extends the stock object by a number of years and estimates the values in the future years of the slots. The future values of the mat, m, harvest.spwn, m.spwn, harvest and all *.wt slots are set to the mean (geometric or arithmetic) value of the last user defined number of years. The future values of the discards.n and landings.n slots are returned only as estimated proportions, not actual numbers. The proportions are calculated as the mean proportion of the sum of the landings and discard numbers over wts.years. The remaining slots (stock.n, catch.n, stock, catch, discards, landings) are set to NA. Not yet implemented for multiple seasons or areas.

See also

Examples

# load ple4
data(ple4)
# stf with default settings
ple4.stf <- stf(ple4, nyears=3, wts.nyears=4, fbar.nyears=3)