FLAssess class and methods

FLAssess is the basic structure for age-based stock assessment. It provides a standard class for data input, diagnostic inspection and stock status estimation; either for use within a working group setting or as part of a formal Management Strategy Evaluation (MSE).

The FLAssess class can be extended to create specific implementations of assessment methods e.g. FLICA, FLSURBA, FLXSA, providing a common interface for all assessmnet methods. For example, within ICES there are two main stock assessment methods, ICA for pelagic and XSA for demersal stocks. However, differences between the methods are mainly artefacts of how they were independently developed rather than methodological. By incorporating such methods in a common class this problem will hopefully be avoided in the future.

FLAssess also incorporates methods for performing virtual population analysis (VPA) and stock projection.

In common with other FLR classes, all of the slots of the FLAssess class have accessor and replacement methods. Users are encouraged to use these instead of the @ slot accessor.

Objects from the Class

Objects can be created by calls to the methods VPA or SepVPA. They also can be created by calls of the form new("FLAssess", ...).

Slots

name

A character string give name of stock.

desc

Whatever you want.

range

A named vector specifying the range of the object.

call

Call that generated the object.

catch.n

An FLQuant that contains estimated catch numbers-at-age.

stock.n

An FLQuant that contains estimated stock numbers-at-age.

harvest

An FLQuant that contains estimated fishing mortlaity-at-age.

index.name

A vector containing names for the FLIndexs

index.range

A list containing ranges for the FLIndexs

index

A list of FLQuants, corresponding to the index values used when fitting.

index.hat

A list of FLQuants, corresponding to the fitted index values.

index.res

A list of FLQuants, corresponding to the index residuals.

index.var

A list of FLQuants, corresponding to the variances of index values in fitting.

Methods

%Summarise contents:
summary:

Returns a list, with a vector for each dimension in the object, that contains the levels of each dimension.

show:

Prints the given object.

plot:

Plots stock.n against index

+:

Updates a FLStock with stock.n and harvest estimates from an FLAssess object

merge:

Updates a FLStock with stock.n and harvest estimates from an FLAssess object

update:

Updates a FLStock with stock.n, catch.n & harvest estimates

%Stock Assessment methods:
SepVPA:

Separable VPA

VPA:

Virtual Population Analysis

% \item{\link{ASPM}:}{Age Structured Production Model}

Author

FLR Team

See also

Examples

   # see the Methods help pages for specific examples